Refactor Options: Separate into multiple pages, and add sidebar menu
