+++ src/ui/about.html
@@ -1,18 +1,46 @@
-<!doctype html>
-<html lang="en">
+<!DOCTYPE html>
+<html>
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Copy as Markdown</title>
-    <link rel="stylesheet" href="../vendor/bulma.css">
+  <title>Copy as Markdown - About</title>
+  <meta charset="utf-8" />
+  <link rel="stylesheet" href="../vendor/bulma.css">
+  <link rel="stylesheet" href="./style.css">
 </head>
-<body class="container section content is-max-desktop">
-<h1>Copy as Markdown</h1>
+<body>
+<div class="container section is-max-desktop">
+  <h1 class="title">Copy as Markdown</h1>
+  <div id="flash-error" class="notification is-danger is-hidden">
+    <p></p>
+  </div>
+  <div class="columns">
+    <div class="column is-narrow" style="width: 200px">
+      <aside class="menu">
+        <p class="menu-label">General</p>
+        <ul class="menu-list">
+          <li><a href="options.html">Markdown Style</a></li>
+          <li><a href="options-permissions.html">Permissions</a></li>
+        </ul>
+        <p class="menu-label">Custom Formats</p>
+        <ul class="menu-list">
+          <li><a href="custom-format.html?slot=1">Custom Format 1</a></li>
+          <li><a href="custom-format.html?slot=2">Custom Format 2</a></li>
+          <li><a href="custom-format.html?slot=3">Custom Format 3</a></li>
+          <li><a href="custom-format.html?slot=4">Custom Format 4</a></li>
+          <li><a href="custom-format.html?slot=5">Custom Format 5</a></li>
+          <li><a href="custom-format-help.html">Help & Examples</a></li>
+        </ul>
+        <p class="menu-label">Others</p>
+        <ul class="menu-list">
+          <li><a href="about.html" class="is-active">About</a></li>
+        </ul>
+      </aside>
+    </div>
+    <div class="column">
+      <div class="box content">
+<h2>About</h2>
 <p>Homepage: <a href="https://github.com/yorkxin/copy-as-markdown">https://github.com/yorkxin/copy-as-markdown</a></p>
 
-<details><summary>License</summary><pre>
+<details><summary>License</summary><pre class="is-overflow-x-scroll">
 The MIT License
 
 Copyright (c) 2012-2024 Yu-Cheng Chuang
@@ -29,7 +57,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 <h3>Bulma</h3>
 <p>Homepage: <a href="https://github.com/jgthms/bulma">https://github.com/jgthms/bulma</a></p>
 
-<details><summary>License</summary><pre>
+<details><summary>License</summary><pre class="is-overflow-x-scroll">
 The MIT License (MIT)
 
 Copyright (c) 2023 Jeremy Thomas
@@ -55,7 +83,7 @@ THE SOFTWARE.
 
 <h3>Turndown.js</h3>
 <p>Homepage: <a href="https://github.com/mixmark-io/turndown">https://github.com/mixmark-io/turndown</a> </p>
-<details><summary>License</summary><pre>
+<details><summary>License</summary><pre class="is-overflow-x-scroll">
 MIT License
 
 Copyright (c) 2017 Dom Christie
@@ -83,7 +111,7 @@ SOFTWARE.
 
 <p>Homepage: <a href="https://github.com/mozilla/webextension-polyfill">https://github.com/mozilla/webextension-polyfill</a></p>
 
-<details><summary>License</summary><pre>
+<details><summary>License</summary><pre class="is-overflow-x-scroll">
 Mozilla Public License Version 2.0
 ==================================
